% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{add4ci}
\alias{add4ci}
\title{Computes the Agresti-Coull `add 4' CI for x successes out of n trials with confidence coefficient conflev. Adds 2 successes and 4 trials.}
\usage{
add4ci(x, n, conflev = 0.95)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of trials}

\item{conflev}{Confidence level}
}
\value{
a list with lower bound, point estimate, and upper bound.
}
\description{
Computes the Agresti-Coull `add 4' CI for x successes out of n trials with confidence coefficient conflev. Adds 2 successes and 4 trials.
}
