% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryinfo.R
\docType{data}
\name{countryinfo}
\alias{countryinfo}
\title{countryinfo}
\format{
An object of class \code{data.frame} with 221 rows and 7 columns.
}
\usage{
countryinfo
}
\description{
countryinfo
}
\keyword{datasets}
