% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-pdf-package-documentation.R
\name{create.package.manual}
\alias{create.package.manual}
\title{create package manual}
\usage{
create.package.manual(pack = "hammond")
}
\arguments{
\item{pack}{name of package}
}
\description{
This function calculates combinations for efficient correlation calculations
}
