% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-management.R
\name{dbkill}
\alias{dbkill}
\title{hdb_login}
\usage{
dbkill()
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
