% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{exact_ind_mover_ci}
\alias{exact_ind_mover_ci}
\title{Square-and-Add or MOVER (Method of Variance Estimate Recovery) Confidence
Interval
Computes a (1 - alpha) * 100% confidence interval for the difference in two
independent proportions
Based on exact binomial (Clopper-Pearson) confidence intervals for single proportions
References:
(1) Robert G. Newcombe. Confidence Intervals for Proportions and Related Measures of Effect Size. CRC Press. 2013
(2) Zou, Huang, and Zhang. A note on confiddence interval estimation for a linear function of binomial proportions. Computational Statistics and Data Analysis. 2009}
\usage{
exact_ind_mover_ci(x1, N1, x2, N2, confidence_level = 0.95)
}
\arguments{
\item{x1}{Number of successes in group 1}

\item{N1}{Number of trials in group 1}

\item{x2}{Number of successes in group 2}

\item{N2}{Number of trials in group 2}

\item{confidence_level}{Confidence level}
}
\value{
a list with lower bound, point estimate, and upper bound.
}
\description{
Square-and-Add or MOVER (Method of Variance Estimate Recovery) Confidence
Interval
Computes a (1 - alpha) * 100% confidence interval for the difference in two
independent proportions
Based on exact binomial (Clopper-Pearson) confidence intervals for single proportions
References:
(1) Robert G. Newcombe. Confidence Intervals for Proportions and Related Measures of Effect Size. CRC Press. 2013
(2) Zou, Huang, and Zhang. A note on confiddence interval estimation for a linear function of binomial proportions. Computational Statistics and Data Analysis. 2009
}
