% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factr.R
\name{factr}
\alias{factr}
\title{Factor based on order}
\usage{
factr(x, y = unique(x))
}
\arguments{
\item{x}{array to factorise}

\item{y}{order of factors}
}
\description{
This makes a factor out of an ordered array
}
