% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-change.R
\name{h.pc.change}
\alias{h.pc.change}
\title{Calculate proportional change}
\usage{
h.pc.change(all)
}
\arguments{
\item{all}{the dataframe to be processed}
}
\value{
Returns a dataframe containing the raw and annual growths in GPI
for each country
}
\description{
This function calculates proportional change in GPI for a country
from one year to another.
}
\examples{
change <- iep.pc.change(df)

}
\keyword{analysis-utils}
\keyword{utilities}
