% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{haddcountryinfo}
\alias{haddcountryinfo}
\title{haddcountryinfo}
\usage{
haddcountryinfo(df)
}
\arguments{
\item{countries}{list of countries}

\item{countries}{list of countries}
}
\description{
This function adds country specific information to a dataframe by matching countries to country codes.
}
\examples{
#need 4 column data frame, geocode, variablename, year, value
hcountryinfo(hcountryexampledata)
hcountry_info

This function calculates correlations between variables

#need 4 column data frame, geocode, variablename, year, value

}
