% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclip.R
\name{hclip}
\alias{hclip}
\title{Save object to clipboard}
\usage{
hclip(obj, size = 4096)
}
\arguments{
\item{obj}{the object to be written to the file}

\item{size}{optional argument used for clipping very large objects}
}
\description{
This function save an object to a clipboard.
}
\examples{
df <- data.frame(a = I("a \\ quote"), b = pi)
hclip(df)

}
\keyword{internal}
\keyword{review}
\keyword{utilities}
