% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcountrycode.R
\name{hcountrycode}
\alias{hcountrycode}
\title{hcountrycode}
\usage{
hcountrycode(x, source_file = whereami::thisfile())
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function replaces country name or code with iso3c country codes. Can also be used in reverse.
}
\examples{
hcountrycode(hcountryexampledata$geocode)

}
