% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcountryexampledata.R
\docType{data}
\name{hcountryexampledata}
\alias{hcountryexampledata}
\title{countryinfo}
\format{An object of class \code{data.frame} with 11934 rows and 5 columns.}
\usage{
hcountryexampledata
}
\description{
countryinfo
}
\keyword{datasets}
