% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcountryinfo.R
\docType{data}
\name{hcountryinfo}
\alias{hcountryinfo}
\title{hcountryinfo}
\format{
An object of class \code{data.frame} with 216 rows and 7 columns.
}
\usage{
hcountryinfo
}
\description{
hcountryinfo
}
\keyword{datasets}
