% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-management.R
\name{hdb_backup}
\alias{hdb_backup}
\title{hdb_backup}
\usage{
hdb_backup(
  host = "192.168.0.98",
  user = "postgres",
  password = "peace123",
  port = 5432
)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function backsup a database, use only if you know what you are doing
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
