% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcorr.R
\name{hdb_corr}
\alias{hdb_corr}
\title{hdb_corr}
\usage{
hdb_corr(df)
}
\arguments{
\item{df}{name of dataframe to use for correlation, needs to be long format 4 column data frame: geocode, variablename, year, value}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, geocode, variablename, year, value
library(hammond)
corr = hcorr(hcountryexampledata)

}
