% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-management.R
\name{hdb_create_db}
\alias{hdb_create_db}
\title{hdb_create_db}
\usage{
hdb_create_db(host, db, user, password)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function creates a database, use only if you know what you are doing
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
