% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{hdb_get}
\alias{hdb_get}
\title{hdb_get}
\usage{
hdb_get(vars)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function retrieves and caches data from any source in the database.
}
\examples{
#need 4 column data frame, geocode, variablename, year, value
hdb_login("192.168.0.98", password = "peace123")
db_get("Perceptions of Criminality Raw")

}
