% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{hdb_get_toc}
\alias{hdb_get_toc}
\title{hdb_get_toc}
\usage{
hdb_get_toc(db = "master")
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function retrieves the Table of Contents from a specified database.
}
\examples{
#need 4 column data frame, geocode, variablename, year, value
hdb_login("192.168.0.98", password = "peace123")
hdb_get_toc()

}
