% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{hdb_login}
\alias{hdb_login}
\title{hdb_login}
\usage{
hdb_login(host = NULL, db = NULL, user = NULL, password = NULL)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function allows access to a database by entering an IP adress and passcode.
}
\examples{
hdb_login("192.168.0.64", db = "countrydata", user = "postgres", password = "peace123")

}
