% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{hdb_search}
\alias{hdb_search}
\title{hdb_search}
\usage{
hdb_search(vars, db = "master")
}
\arguments{
\item{countries}{list of countries

hdb_login("192.168.0.98", password = "peace123")
hdb_search("Criminal)}
}
\description{
This function searches the database and retrieves specified data.
}
