% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-management.R
\name{hdb_update_master}
\alias{hdb_update_master}
\title{hdb_update_master}
\usage{
hdb_update_master()
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
