% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfont.R
\name{hfont}
\alias{hfont}
\title{hfonts}
\usage{
hfont(libname = find.package("hammond"), pkgname = "hammond")
}
\arguments{
\item{df}{}
}
\description{
This function loads hammond fonts
}
\examples{
x = hfont()

}
