% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinterpolate.R
\name{hinterpolate}
\alias{hinterpolate}
\title{hinterpolate}
\usage{
hinterpolate(df)
}
\arguments{
\item{df}{dataframe in iep format}
}
\value{
Returns list with filled in time series, column yhat is the interpolated value.
Please check original value with yhat column to make sure you are happy with the results
}
\description{
This is a wrapper function takes a data frame and fills in
interpolated and extrapolated data for the whole time series
}
\examples{
hinterpolate(hcountryexampledata)

}
\author{
Dave
}
\keyword{imputation}
