% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpack-manual.R
\name{hpack_manual}
\alias{hpack_manual}
\title{create package manual}
\usage{
hpack_manual(pack = "hammond")
}
\arguments{
\item{pack}{name of package}
}
\description{
This function calculates combinations for efficient correlation calculations
}
