% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc-change.R
\name{hpc_change}
\alias{hpc_change}
\title{Calculate proportional change}
\usage{
hpc_change(all)
}
\arguments{
\item{all}{the dataframe to be processed}
}
\value{
Returns a dataframe containing the raw and annual growths in GPI
for each country
hpc_change(hcountryexampledata)
}
\description{
This function calculates proportional change in GPI for a country
from one year to another.
}
\keyword{analysis-utils}
\keyword{utilities}
