% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpopulation.R
\name{hpopulation}
\alias{hpopulation}
\title{hpopulation}
\usage{
hpopulation(df)
}
\arguments{
\item{df}{}
}
\description{
This function appends a column of populations to a dataframe
}
\examples{
x = hpopulation(hammond::hcountryexampledata)

}
