% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{install dependencies}
\usage{
ipak(pkg.name)
}
\description{
easy install for R packages, sourced from
https://gist.github.com/stevenworthington/3178163
}
\examples{
ipak("tidyverse")

}
