% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Monte carlo runs n_simulations and calls the callback function each time with the ... optional args}
\usage{
monte_carlo(n_simulations, callback, ...)
}
\arguments{
\item{n_simulations}{Number of simulations to run}

\item{callback}{Function to call each simulation}
}
\description{
Monte carlo runs n_simulations and calls the callback function each time with the ... optional args
}
