% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytheme.R
\name{mytheme}
\alias{mytheme}
\title{My plot theme}
\usage{
mytheme(title_size = 22, axis_and_legend_title_size = 15, text_size = 12)
}
\arguments{
\item{title_size}{Size of the title text}

\item{axis_and_legend_title_size}{Size of the legend title, axis titles, and subtitle text.}

\item{text_size}{Size of axis and legend text}
}
\value{
a ggplot theme object
}
\description{
My plot theme
}
