% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_calc}
\alias{power_calc}
\title{This function performs a power analysis using the given method}
\usage{
power_calc(
  method,
  effect_sizes = NULL,
  sample_sizes = NULL,
  power = NULL,
  x_reference = NULL,
  n_reference = NULL,
  alpha = 0.05,
  numsims = 1000
)
}
\arguments{
\item{method}{Metho of calculating confidence interval}

\item{effect_sizes}{Effect size(s) to examine}

\item{sample_sizes}{Sample sizes to examine}
}
\value{
a list with lower bound, point estimate, and upper bound.
}
\description{
This function performs a power analysis using the given method
}
