% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_utility.R
\name{round_df}
\alias{round_df}
\title{round_df() rounds the numeric columns of the input data frame to the specified number of digits.}
\usage{
round_df(data, digits)
}
\arguments{
\item{data}{data frame to round}

\item{digits}{Number of digits to round to}
}
\description{
round_df() rounds the numeric columns of the input data frame to the specified number of digits.
}
