% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{scoreci}
\alias{scoreci}
\title{This function computes a confidence interval for a proportion. It is based on inverting the large-sample normal score test for the proportion.}
\usage{
scoreci(x, n, conflev = 0.95)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of trials}

\item{conflev}{Confidence level}
}
\value{
a list with lower bound, point estimate, and upper bound.
}
\description{
This function computes a confidence interval for a proportion. It is based on inverting the large-sample normal score test for the proportion.
}
