% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsplit.R
\name{split}
\alias{split}
\title{Create a two factor list}
\usage{
split(df, factor1, factor2)
}
\arguments{
\item{df}{name of dataframe}

\item{factor1}{first factor to split on}

\item{factor2}{second factor to split on}
}
\description{
Create a two factor list
}
