% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{xy_boxplots}
\alias{xy_boxplots}
\title{Plot x vs. y with boxplots outside of each axis}
\usage{
xy_boxplots(data, x, y, ...)
}
\arguments{
\item{data}{A dataframe with the columns "x" and "y" (numeric) to be plotted.}

\item{x}{A character string "x" with the column name to be plotted on the x-axis}

\item{y}{A character string "y" with the column name to be plotted on the y-axis}

\item{...}{Optional arguments `color` and `color_labels`. `color` must be a factor, `color_labels` must be a character string of the same length as the number of levels in factor `color`}
}
\value{
a ggplot object
}
\description{
Plot x vs. y with boxplots outside of each axis
}
