% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{h.correlate}
\alias{h.correlate}
\title{hcorrelate}
\usage{
h.correlate(df, min.pairs = 20, verbose = TRUE, filter.by.p = FALSE)
}
\arguments{
\item{df}{name of dataframe to use for correlation, needs to be long format 4 column data frame: iso3c, variablename, year, value}

\item{min.pairs}{minimum number of pairs to correlate}

\item{verbose}{enable n and p values reporting, TRUE or FALSE}

\item{filter.by.p}{Do you want to filter for significant p values?}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, iso3c, variablename, year, value
}
