% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammond.R
\docType{package}
\name{hammond-package}
\alias{hammond}
\alias{hammond-package}
\title{hammond: some stuff}
\description{
hammond: some stuff
}
\section{Installation}{

\code{devtools::install_github("david-hammond/hammond")}
}

