% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcountrycode.R
\name{hcountrycode}
\alias{hcountrycode}
\title{hcountrycode}
\usage{
hcountrycode(countries = c("ALG", "ALB", "UKG", "CAN", "USA"))
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, iso3c, variablename, year, value

}
