% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcountryinfo.R
\docType{data}
\name{hcountryinfo}
\alias{hcountryinfo}
\title{countryinfo}
\format{An object of class \code{data.frame} with 233 rows and 7 columns.}
\usage{
hcountryinfo
}
\description{
countryinfo
}
\keyword{datasets}
