% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{hcountryinfo}
\alias{hcountryinfo}
\title{hcountry_info}
\usage{
hcountryinfo(df, host = NULL, password = NULL)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, iso3c, variablename, year, value

}
