% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcountryinfo.R, R/hdb-utils.R
\name{hcountryinfo}
\alias{hcountryinfo}
\title{countryinfo}
\usage{
hcountryinfo(df)

hcountryinfo(df)
}
\arguments{
\item{countries}{list of countries}

\item{countries}{list of countries}
}
\description{
countryinfo

This function calculates correlations between variables
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

hcountry_info

This function calculates correlations between variables

#need 4 column data frame, geocode, variablename, year, value

}
