% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{hdb_connect}
\alias{hdb_connect}
\title{hdb_connect}
\usage{
hdb_connect(db = "postgres", host = my.name <- readline(prompt =
  "Enter database ip address: "), port = 5432, user = "postgres",
  password = my.name <- readline(prompt = "Enter database password: "))
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, iso3c, variablename, year, value

}
