% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-management.R
\name{hdb_create_db}
\alias{hdb_create_db}
\title{hdb_create_db}
\usage{
hdb_create_db(db)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
