% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-utils.R
\name{hdb_get_toc}
\alias{hdb_get_toc}
\title{hdb_get_toc}
\usage{
hdb_get_toc(db = "master", host = NULL, password = NULL)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
