% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb-management.R
\name{hdbkill}
\alias{hdbkill}
\title{hdb_kill}
\usage{
hdbkill()
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function kills all connections to the database, use as a last resort if you get a db connection error
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
