% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinterpolate.R
\name{hinterpolate}
\alias{hinterpolate}
\title{hinterpolate}
\usage{
hinterpolate(df)
}
\arguments{
\item{countries}{list of countries}
}
\description{
This function calculates correlations between variables
}
\examples{
#need 4 column data frame, geocode, variablename, year, value

}
